Moved `get_image` from Models to Utils for DRYness
